Moved hooks to seperated files
