+++ src/reanimated2/hooks/index.ts
@@ -1,1 +1,8 @@
 export * from './Hooks';
+export * from './useSharedValue';
+export * from './useAnimatedStyle';
+export * from './useAnimatedGestureHandler';
+export * from './useAnimatedReaction';
+export * from './useAnimatedRef';
+export * from './useAnimatedScrollHandler';
+export * from './useDerivedValue';
