+++ src/reanimated2/hooks/useAnimatedScrollHandler.ts
@@ -1,0 +1,80 @@
+import { useEffect, useRef } from 'react';
+import { makeRemote } from '../core';
+import { areDependenciesEqual, buildDependencies, useEvent } from './utils';
+
+export function useAnimatedScrollHandler(handlers, dependencies) {
+  const initRef = useRef(null);
+  if (initRef.current === null) {
+    initRef.current = {
+      context: makeRemote({}),
+      savedDependencies: [],
+    };
+  }
+
+  useEffect(() => {
+    return () => {
+      initRef.current = null;
+    };
+  }, []);
+
+  const { context, savedDependencies } = initRef.current;
+
+  dependencies = buildDependencies(dependencies, handlers);
+
+  const dependenciesDiffer = !areDependenciesEqual(
+    dependencies,
+    savedDependencies
+  );
+  initRef.current.savedDependencies = dependencies;
+
+  // build event subscription array
+  const subscribeForEvents = ['onScroll'];
+  if (handlers.onBeginDrag !== undefined) {
+    subscribeForEvents.push('onScrollBeginDrag');
+  }
+  if (handlers.onEndDrag !== undefined) {
+    subscribeForEvents.push('onScrollEndDrag');
+  }
+  if (handlers.onMomentumBegin !== undefined) {
+    subscribeForEvents.push('onMomentumScrollBegin');
+  }
+  if (handlers.onMomentumEnd !== undefined) {
+    subscribeForEvents.push('onMomentumScrollEnd');
+  }
+
+  return useEvent(
+    (event) => {
+      'worklet';
+      const {
+        onScroll,
+        onBeginDrag,
+        onEndDrag,
+        onMomentumBegin,
+        onMomentumEnd,
+      } = handlers;
+      if (event.eventName.endsWith('onScroll')) {
+        if (onScroll) {
+          onScroll(event, context);
+        } else if (typeof handlers === 'function') {
+          handlers(event, context);
+        }
+      } else if (onBeginDrag && event.eventName.endsWith('onScrollBeginDrag')) {
+        onBeginDrag(event, context);
+      } else if (onEndDrag && event.eventName.endsWith('onScrollEndDrag')) {
+        onEndDrag(event, context);
+      } else if (
+        onMomentumBegin &&
+        event.eventName.endsWith('onMomentumScrollBegin')
+      ) {
+        onMomentumBegin(event, context);
+      } else if (
+        onMomentumEnd &&
+        event.eventName.endsWith('onMomentumScrollEnd')
+      ) {
+        onMomentumEnd(event, context);
+      }
+    },
+    subscribeForEvents,
+    dependenciesDiffer
+  );
+}
